requirements.txt
Jinja2==3.0.2
kafka-python==2.0.2
redis==4.1.2
requests==2.26.0
python-dotenv==0.19.2
certifi
confluent-kafka[avro,json,protobuf]>=1.4.2
